+++ sdk/formrecognizer/azure-ai-formrecognizer/azure/ai/formrecognizer/_form_training_client.py
@@ -29,7 +29,7 @@ from ._generated.models import (
 from ._helpers import error_map, get_authentication_policy, POLLING_INTERVAL, TransportWrapper
 from ._models import (
     CustomFormModelInfo,
-    AccountProperties,
+    FormRecognizerAccountProperties,
     CustomFormModel
 )
 from ._polling import TrainingPolling, CopyPolling
@@ -231,12 +231,12 @@ class FormTrainingClient(object):
 
     @distributed_trace
     def get_account_properties(self, **kwargs):
-        # type: (Any) -> AccountProperties
+        # type: (Any) -> FormRecognizerAccountProperties
         """Get information about the models on the form recognizer account.
 
         :return: Summary of models on account - custom model count,
             custom model limit.
-        :rtype: ~azure.ai.formrecognizer.AccountProperties
+        :rtype: ~azure.ai.formrecognizer.FormRecognizerAccountProperties
         :raises ~azure.core.exceptions.HttpResponseError:
 
         .. admonition:: Example:
@@ -249,7 +249,7 @@ class FormTrainingClient(object):
                 :caption: Get properties for the form recognizer account.
         """
         response = self._client.get_custom_models(error_map=error_map, **kwargs)
-        return AccountProperties._from_generated(response.summary)
+        return FormRecognizerAccountProperties._from_generated(response.summary)
 
     @distributed_trace
     def get_custom_model(self, model_id, **kwargs):